fix_bugs

+++ leaven/src/lean_manager.py
@@ -35,46 +35,50 @@ def get_lean(version='3.51.1', remove_installation=True):
     bin_path = parent_path / 'elan' / 'bin' / 'elan'
     if os.path.exists(Path(__file__).resolve().parent.parent / 'elan' / 'toolchains' / f'leanprover-community--lean---{version}') or \
         os.path.exists(Path(__file__).resolve().parent.parent / 'elan' / 'toolchains' / f'leanprover--lean4---{version}'):
-        print(f'Lean {version} exists')
-    else:
-        if not check_connection(f'https://github.com/leanprover-community/lean/releases/download/v{version}/lean-{version}-windows.zip'):
-            raise Exception(f'Fail to connect to the URL: https://github.com/leanprover-community/lean/releases/download/v{version}/lean-{version}-windows.zip')
-        if version.startswith('3'):
-            if platform.system() == 'Windows':
-                url = f'https://github.com/leanprover-community/lean/releases/download/v{version}/lean-{version}-windows.zip'
-            elif platform.system() == 'Linux':
-                if platform.machine() == 'x86_64':
-                    url = f'https://github.com/leanprover-community/lean/releases/download/v{version}/lean-{version}-linux.tar.gz'
-                elif platform.machine() == 'AArch64':
-                    url = f'https://github.com/leanprover-community/lean/releases/download/v{version}/lean-{version}-linux_aarch64.tar.gz'
-                else:
-                    raise ValueError('Unknown platform')
-            elif platform.system() == 'Darwin':
-                url = f'https://github.com/leanprover-community/lean/releases/download/v{version}/lean-{version}-darwin.zip'
+        return
+    if not check_connection(f'https://github.com/leanprover-community/lean/releases/download/v{version}/lean-{version}-windows.zip'):
+        raise Exception(f'Fail to connect to the URL: https://github.com/leanprover-community/lean/releases/download/v{version}/lean-{version}-windows.zip')
+    if version.startswith('3'):
+        if platform.system() == 'Windows':
+            url = f'https://github.com/leanprover-community/lean/releases/download/v{version}/lean-{version}-windows.zip'
+        elif platform.system() == 'Linux':
+            if platform.machine() == 'x86_64':
+                url = f'https://github.com/leanprover-community/lean/releases/download/v{version}/lean-{version}-linux.tar.gz'
+            elif platform.machine() == 'AArch64':
+                url = f'https://github.com/leanprover-community/lean/releases/download/v{version}/lean-{version}-linux_aarch64.tar.gz'
             else:
                 raise ValueError('Unknown platform')
+        elif platform.system() == 'Darwin':
+            url = f'https://github.com/leanprover-community/lean/releases/download/v{version}/lean-{version}-darwin.zip'
         else:
-            if platform.system() == 'Windows':
-                url = f'https://github.com/leanprover/lean4/releases/download/v{version}/lean-{version}-windows.zip'
-            elif platform.system() == 'Linux':
-                if platform.machine() == 'x86':
-                    url = f'https://github.com/leanprover/lean4/releases/download/v{version}/lean-{version}-linux.tar.gz'
-                elif platform.machine() == 'AArch64':
-                    url = f'https://github.com/leanprover/lean4/releases/download/v{version}/lean-{version}-linux_aarch64.tar.gz'
-            elif platform.system() == 'Darwin':
-                url = f'https://github.com/leanprover/lean4/releases/download/v{version}/lean-{version}-darwin.zip'
-            else:
-                raise ValueError('Unknown platform')
-        file = Path(__file__).resolve().parent.parent / os.path.basename(url)
-        with urllib.request.urlopen(url) as response, open(file, 'wb') as out_file:
-            shutil.copyfileobj(response, out_file)
-        extracted_file = extract_file(file)
-        shutil.move(extracted_file, toolchain_path)
-        if remove_installation:
-            os.remove(file)
+            raise ValueError('Unknown platform')
+    else:
+        if platform.system() == 'Windows':
+            url = f'https://github.com/leanprover/lean4/releases/download/v{version}/lean-{version}-windows.zip'
+        elif platform.system() == 'Linux':
+            if platform.machine() == 'x86':
+                url = f'https://github.com/leanprover/lean4/releases/download/v{version}/lean-{version}-linux.tar.gz'
+            elif platform.machine() == 'AArch64':
+                url = f'https://github.com/leanprover/lean4/releases/download/v{version}/lean-{version}-linux_aarch64.tar.gz'
+        elif platform.system() == 'Darwin':
+            url = f'https://github.com/leanprover/lean4/releases/download/v{version}/lean-{version}-darwin.zip'
+        else:
+            raise ValueError('Unknown platform')
+    file = Path(__file__).resolve().parent.parent / os.path.basename(url)
+    with urllib.request.urlopen(url) as response, open(file, 'wb') as out_file:
+        shutil.copyfileobj(response, out_file)
+    extracted_file = extract_file(file)
+    shutil.move(extracted_file, toolchain_path)
+    if remove_installation:
+        os.remove(file)
     assert os.system(f'{bin_path} toolchain link {toolchain} {toolchain_path}') == 0
 
-def download_mathlib_datasets(remove_installation=True):
+def get_mathlib_datasets(remove_installation=True):
+    if os.path.exists(Path(__file__).resolve().parent.parent / '_target') and \
+        os.path.exists(Path(__file__).resolve().parent.parent / 'proving_env_lib.bak') and \
+            os.path.exists(Path(__file__).resolve().parent.parent / 'proving_env_lib.dat') and \
+                os.path.exists(Path(__file__).resolve().parent.parent / 'proving_env_lib.dir'):
+        return
     url = 'https://github.com/xinhjBrant/leaven/releases/download/release/mathlib_dataset.tar.gz'
     file_name = Path(__file__).resolve().parent.parent / os.path.basename(url)
     with urllib.request.urlopen(url) as response, open(file_name, 'wb') as out_file:
